[Major] moving bot class into its own file
